[Still following the course] Added new features
Programmed the movement of the rocket:
Now through the space bar the rocket goes up.
Through A and D the rocket goes respectively to the left and to the right.

Constraints have been added to the movement of the rocket:
Now the rocket cannot rotate on the X and Y axis.
Now the rocket cannot move on the Z axis

Added method to manage gravity change when changing levels aka planet