Various additions
Now when the player hits an obstacle or reaches the destination a message is printed on the screen.

Created the second level of the game, now if the player reaches the destination the level next to the one just played is loaded.

I created an object that will manage the entire UI of the game (now it only takes care of printing messages on the screen)