First pre-release commit
Additions:
Now when the player hits an object other than the base of the launch pad / landing pad it's game over! The level reloads and the player will be on the launch pad.
Added screen text when hitting an object or when reaching the landing pad.
Added a camera that follows the player wherever he goes.
Created the "Earth" level and the "Mars" level, the latter is only a copy of the "Earth" level.
Implemented a method to change the level gravity when it (the level) changes.

Bugfix:
After adding the camera due to a problem during the game you could run into heavy stuttering of the rocket as it moved, I solved it by adding the FixedUpdate method and commenting the Update method, also I added the "Interpolate" property to the Rigidbody component of the rocket.
ReadMe file updated.